+++ packages/react/src/ThemeContext.ts
@@ -1,0 +1,7 @@
+import React from 'react';
+import { ThemeContextShape } from './types';
+
+export default React.createContext<ThemeContextShape>({
+  changeTheme() {},
+  theme: '',
+});