+++ packages/react/src/index.ts
@@ -5,6 +5,8 @@
 
 import DirectionContext from './DirectionContext';
 import DirectionProvider from './DirectionProvider';
+import ThemeContext from './ThemeContext';
+import ThemeProvider from './ThemeProvider';
 import useStylesFactory from './useStylesFactory';
 import useThemeFactory from './useThemeFactory';
 import withStylesFactory from './withStylesFactory';
@@ -15,6 +17,8 @@ export * from './types';
 export {
   DirectionContext,
   DirectionProvider,
+  ThemeContext,
+  ThemeProvider,
   useStylesFactory,
   useThemeFactory,
   withStylesFactory,